\name{paurelia}
\alias{paurelia}
\docType{data}
\title{
Paramecium abundance time series
}
\description{
Paramecium abundance time series
}
\usage{data(paurelia)}
\format{
  The format is:
 num [1:20] 2 NA 17 29 39 63 185 258 267 392 ...
}
\details{
Paramecium abundance time series with missing value.
}
\source{
FIXME
}
\references{
FIXME
}
\examples{
data(paurelia)
paurelia
plot(paurelia)
}
\keyword{datasets}
