\name{pva-methods}
\docType{methods}
\alias{summary,pva-method}
\alias{coef,pva-method}
\alias{confint,pva-method}
\alias{vcov,pva-method}
\title{Methods for 'pva' objects}
\description{
\code{coef}, \code{vcov}, \code{confint},
and \code{summary} methods for 'pva' objects.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "pva")}}{
Methods for S4 objects of class 'pva'.
}
}}
\keyword{methods}
