\name{songsparrow}
\alias{songsparrow}
\docType{data}
\title{
Songsparrow abundance time series
}
\description{
Songsparrow abundance time series
}
\usage{data(songsparrow)}
\format{
  The format is:
 num [1:30] 18 10 9 14 17 14 5 10 9 5 ...
}
\details{
Songsparrow abundance time series
}
\source{
FIXME
}
\references{
FIXME
}
\examples{
data(songsparrow)
songsparrow
plot(songsparrow)
}
\keyword{datasets}
