\name{paurelia}
\alias{paurelia}
\docType{data}
\title{
Paramecium abundance time series
}
\description{
\emph{Paramecium aurelia} abundance time series
}
\usage{data(paurelia)}
\format{
  The format is:
 num [1:20] 2 NA 17 29 39 63 185 258 267 392 ...
}
\details{
\emph{Paramecium aurelia} abundance time series with a missing value.
}
\source{
Gause (1934: Appendix I, Table 3)
}
\references{
Gause, G.F. (1934). The Struggle for Existence. Williams & Wilkins, Baltimore.
}
\examples{
data(paurelia)
paurelia
plot(paurelia)
}
\keyword{datasets}
