\name{pva-methods}
\docType{methods}
\alias{show-methods}
\alias{show,pva-method}
\alias{as.mcmc.list-methods}
\alias{as.mcmc.list,pva-method}
\alias{coef-methods}
\alias{coef,pva-method}
\alias{confint-methods}
\alias{confint,pva-method}
\alias{vcov-methods}
\alias{vcov,pva-method}

\title{Methods for 'pva' objects}
\description{
\code{coef}, \code{vcov}, \code{confint},
\code{show} and \code{as.mcmc.list} coercion
methods for 'pva' objects.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "pva")}}{
Methods for S4 objects of class 'pva'.
}
}}
\keyword{methods}
