\name{redstart}
\alias{redstart}
\docType{data}
\title{
redstart abundance time series
}
\description{
redstart abundance time series
}
\usage{data(redstart)}
\format{
  The format is:
 num [1:30] 18 10 9 14 17 14 5 10 9 5 ...
}
\details{
redstart abundance time series
}
\source{
Peter Arcese kindly provided the song sparrow population counts data.
}
\references{

FIXME --- KHURRAM --- to check references for redstart

Saether, B-E., Engen, S., Lande, R., Acrese, P. and Smith, J.N.M., 2000. 
Estimating the time to extinction in an island population 
of song sparrows. Proc. R. Soc. Lond. B 267, 621-626.

Nadeem, K., Lele S. R., 2012.  Likelihood based population 
viability analysis in the presence of 
observation error. Oikos. doi: 10.1111/j.1600-0706.2011.20010.x


}
\examples{
data(redstart)
redstart
plot(redstart)
}
\keyword{datasets}
