\name{songsparrow}
\alias{songsparrow}
\docType{data}
\title{
Songsparrow abundance time series
}
\description{
Songsparrow abundance time series
}
\usage{data(songsparrow)}
\format{
  The format is:
 num [1:30] 18 10 9 14 17 14 5 10 9 5 ...
}
\details{
Songsparrow abundance time series
}
\source{
Peter Arcese kindly provided the song sparrow population counts data.
}
\references{

Saether, B. et al. 2000. Estimating the time to extinction in an island population 
of song sparrows. Proc. R. Soc. Lond. B 267, 621-626.

Nadeem, K., Lele S. R., 2012.  Likelihood based population viability analysis in the presence of 
observation error. Oikos. doi: 10.1111/j.1600-0706.2011.20010.x


}
\examples{
data(songsparrow)
songsparrow
plot(songsparrow)
}
\keyword{datasets}
